feat: refactor index-page into components
